Bug: fix bug with server error 308
The error was caused because I was making the request from client using the whole route http://localhost:5000/api/predict. When I changed it to simply /api/predict it worked fine and CORS wasn't a problem. The problem was caused by chromium blocking requests from different domains. And it thought this request was being done from an outside domain.